First draft of runchart reusing scatter and using mvvm pattern
