+++ client/plots/runChart.js
@@ -1,1 +1,1 @@
-export * from './runChart'
+export * from './runchart/runchart.js'
