+++ client/plots/scatter/viewmodel/scatterTooltip.ts
@@ -1,9 +1,9 @@
 import { mclass } from '#shared/common.js'
-import { shapesArray, shapeSelector } from '../../../dom/shapes.js'
+import { shapesArray } from '../../../dom/shapes.js'
 import { roundValueAuto } from '#shared/roundValue.js'
 import { getDateStrFromNumber } from '#shared/terms.js'
 import { rgb } from 'd3-color'
-import { Scatter } from '../Scatter.js'
+import type { Scatter } from '../scatter.js'
 
 export class ScatterTooltip {
 	scatter: Scatter
@@ -232,7 +232,7 @@ export class ScatterTooltip {
 							.append('td')
 							.append('button')
 							.text('Lollipop')
-							.on('click', async e => {
+							.on('click', async () => {
 								await this.scatter.interactivity.openLollipop(label)
 								this.scatter.dom.tip.hide()
 							})
@@ -249,7 +249,7 @@ export class ScatterTooltip {
 					.attr('stroke', '#aaa')
 					.attr('transform', 'translate(0, -2) scale(0.5)')
 				const text = g.append('text').attr('x', 12).attr('y', 6).attr('font-size', '0.9em')
-				const span2 = text.append('tspan').text(node.value).attr('fill', fontColor)
+				text.append('tspan').text(node.value).attr('fill', fontColor)
 			} else td.style('padding', '2px').text(`${node.value}`)
 		}
 
@@ -271,20 +271,20 @@ export class ScatterTooltip {
 						.append('td')
 						.append('button')
 						.text('Sample view')
-						.on('click', e => this.scatter.interactivity.openSampleView(sample))
+						.on('click', () => this.scatter.interactivity.openSampleView(sample))
 					if (hasDiscoPlot)
 						row
 							.append('td')
 							.append('button')
 							.text('Disco')
-							.on('click', async e => this.scatter.interactivity.openDiscoPlot(sample))
+							.on('click', async () => this.scatter.interactivity.openDiscoPlot(sample))
 
 					if (hasMetArrayPlot)
 						row
 							.append('td')
 							.append('button')
 							.text('Met Array')
-							.on('click', async e => this.scatter.interactivity.openMetArray(sample))
+							.on('click', async () => this.scatter.interactivity.openMetArray(sample))
 				}
 			}
 		}
